Extracted base64 persistence functionality
