+++ modules/core/src/persistence.cpp
@@ -41,26 +41,13 @@
 //M*/
 
 #include "precomp.hpp"
-
+#include "persistence.hpp"
 #include <ctype.h>
 #include <deque>
 #include <sstream>
 #include <string>
 #include <iterator>
 
-#define USE_ZLIB 1
-
-#if USE_ZLIB
-#  ifndef _LFS64_LARGEFILE
-#    define _LFS64_LARGEFILE 0
-#  endif
-#  ifndef _FILE_OFFSET_BITS
-#    define _FILE_OFFSET_BITS 0
-#  endif
-#  include <zlib.h>
-#else
-typedef void* gzFile;
-#endif
 
 /****************************************************************************************\
 *                            Common macros and type definitions                          *
@@ -177,16 +164,6 @@ cv::String cv::FileStorage::getDefaultObjectName(const cv::String& _filename)
     return String(name);
 }
 
-typedef struct CvGenericHash
-{
-    CV_SET_FIELDS()
-    int tab_size;
-    void** table;
-}
-CvGenericHash;
-
-typedef CvGenericHash CvStringHash;
-
 typedef struct CvFileMapNode
 {
     CvFileNode value;
@@ -204,20 +181,6 @@ typedef struct CvXMLStackRecord
 }
 CvXMLStackRecord;
 
-namespace base64
-{
-    class Base64Writer;
-
-    namespace fs
-    {
-        enum State
-        {
-            Uncertain,
-            NotUse,
-            InUse,
-        };
-    }
-}
 
 #define CV_XML_OPENING_TAG 1
 #define CV_XML_CLOSING_TAG 2
@@ -225,173 +188,8 @@ namespace base64
 #define CV_XML_HEADER_TAG 4
 #define CV_XML_DIRECTIVE_TAG 5
 
-//typedef void (*CvParse)( struct CvFileStorage* fs );
-typedef void (*CvStartWriteStruct)( struct CvFileStorage* fs, const char* key,
-                                    int struct_flags, const char* type_name );
-typedef void (*CvEndWriteStruct)( struct CvFileStorage* fs );
-typedef void (*CvWriteInt)( struct CvFileStorage* fs, const char* key, int value );
-typedef void (*CvWriteReal)( struct CvFileStorage* fs, const char* key, double value );
-typedef void (*CvWriteString)( struct CvFileStorage* fs, const char* key,
-                               const char* value, int quote );
-typedef void (*CvWriteComment)( struct CvFileStorage* fs, const char* comment, int eol_comment );
-typedef void (*CvStartNextStream)( struct CvFileStorage* fs );
 
-typedef struct CvFileStorage
-{
-    int flags;
-    int fmt;
-    int write_mode;
-    int is_first;
-    CvMemStorage* memstorage;
-    CvMemStorage* dststorage;
-    CvMemStorage* strstorage;
-    CvStringHash* str_hash;
-    CvSeq* roots;
-    CvSeq* write_stack;
-    int struct_indent;
-    int struct_flags;
-    CvString struct_tag;
-    int space;
-    char* filename;
-    FILE* file;
-    gzFile gzfile;
-    char* buffer;
-    char* buffer_start;
-    char* buffer_end;
-    int wrap_margin;
-    int lineno;
-    int dummy_eof;
-    const char* errmsg;
-    char errmsgbuf[128];
-
-    CvStartWriteStruct start_write_struct;
-    CvEndWriteStruct end_write_struct;
-    CvWriteInt write_int;
-    CvWriteReal write_real;
-    CvWriteString write_string;
-    CvWriteComment write_comment;
-    CvStartNextStream start_next_stream;
-
-    const char* strbuf;
-    size_t strbufsize, strbufpos;
-    std::deque<char>* outbuf;
-
-    base64::Base64Writer * base64_writer;
-    bool is_default_using_base64;
-    base64::fs::State state_of_writing_base64;  /**< used in WriteRawData only */
-
-    bool is_write_struct_delayed;
-    char* delayed_struct_key;
-    int   delayed_struct_flags;
-    char* delayed_type_name;
-
-    bool is_opened;
-}
-CvFileStorage;
-
-namespace base64
-{
-    static const size_t HEADER_SIZE         = 24U;
-    static const size_t ENCODED_HEADER_SIZE = 32U;
-
-    /* base64 */
-
-    typedef uchar uint8_t;
-
-    extern uint8_t const base64_padding;
-    extern uint8_t const base64_mapping[65];
-    extern uint8_t const base64_demapping[127];
-
-    size_t base64_encode(uint8_t const * src, uint8_t * dst, size_t off,      size_t cnt);
-    size_t base64_encode(   char const * src,    char * dst, size_t off = 0U, size_t cnt = 0U);
-
-    size_t base64_decode(uint8_t const * src, uint8_t * dst, size_t off,      size_t cnt);
-    size_t base64_decode(   char const * src,    char * dst, size_t off = 0U, size_t cnt = 0U);
-
-    bool   base64_valid (uint8_t const * src, size_t off,      size_t cnt);
-    bool   base64_valid (   char const * src, size_t off = 0U, size_t cnt = 0U);
-
-    size_t base64_encode_buffer_size(size_t cnt, bool is_end_with_zero = true);
-
-    size_t base64_decode_buffer_size(size_t cnt, bool is_end_with_zero = true);
-    size_t base64_decode_buffer_size(size_t cnt, char  const * src, bool is_end_with_zero = true);
-    size_t base64_decode_buffer_size(size_t cnt, uchar const * src, bool is_end_with_zero = true);
-
-    /* binary */
-
-    template<typename _uint_t>      inline size_t to_binary(_uint_t       val, uchar * cur);
-    template<>                      inline size_t to_binary(double        val, uchar * cur);
-    template<>                      inline size_t to_binary(float         val, uchar * cur);
-    template<typename _primitive_t> inline size_t to_binary(uchar const * val, uchar * cur);
-
-    template<typename _uint_t>      inline size_t binary_to(uchar const * cur, _uint_t & val);
-    template<>                      inline size_t binary_to(uchar const * cur, double  & val);
-    template<>                      inline size_t binary_to(uchar const * cur, float   & val);
-    template<typename _primitive_t> inline size_t binary_to(uchar const * cur, uchar   * val);
-
-    class RawDataToBinaryConvertor;
-
-    class BinaryToCvSeqConvertor;
-
-    /* class */
-
-    class Base64ContextParser
-    {
-    public:
-        explicit Base64ContextParser(uchar * buffer, size_t size);
-        ~Base64ContextParser();
-        Base64ContextParser & read(const uchar * beg, const uchar * end);
-        bool flush();
-    private:
-        static const size_t BUFFER_LEN = 120U;
-        uchar * dst_cur;
-        uchar * dst_end;
-        std::vector<uchar> base64_buffer;
-        uchar * src_beg;
-        uchar * src_cur;
-        uchar * src_end;
-        std::vector<uchar> binary_buffer;
-    };
-
-    class Base64ContextEmitter;
-
-    class Base64Writer
-    {
-    public:
-        Base64Writer(::CvFileStorage * fs);
-        ~Base64Writer();
-        void write(const void* _data, size_t len, const char* dt);
-        template<typename _to_binary_convertor_t> void write(_to_binary_convertor_t & convertor, const char* dt);
-
-    private:
-        void check_dt(const char* dt);
-
-    private:
-        // disable copy and assignment
-        Base64Writer(const Base64Writer &);
-        Base64Writer & operator=(const Base64Writer &);
-
-    private:
-
-        Base64ContextEmitter * emitter;
-        std::string data_type_string;
-    };
-
-    /* other */
-
-    std::string make_base64_header(const char * dt);
-
-    bool read_base64_header(std::vector<char> const & header, std::string & dt);
-
-    void make_seq(void * binary_data, int elem_cnt, const char * dt, CvSeq & seq);
-
-    /* sample */
-
-    void cvWriteRawDataBase64(::CvFileStorage* fs, const void* _data, int len, const char* dt);
-}
-
-
-static void icvPuts( CvFileStorage* fs, const char* str )
+void icvPuts( CvFileStorage* fs, const char* str )
 {
     if( fs->outbuf )
         std::copy(str, str + strlen(str), std::back_inserter(*fs->outbuf));
@@ -405,7 +203,7 @@ static void icvPuts( CvFileStorage* fs, const char* str )
         CV_Error( CV_StsError, "The storage is not opened" );
 }
 
-static char* icvGets( CvFileStorage* fs, char* str, int maxCount )
+char* icvGets( CvFileStorage* fs, char* str, int maxCount )
 {
     if( fs->strbuf )
     {
@@ -435,7 +233,7 @@ static char* icvGets( CvFileStorage* fs, char* str, int maxCount )
     return 0;
 }
 
-static int icvEof( CvFileStorage* fs )
+int icvEof( CvFileStorage* fs )
 {
     if( fs->strbuf )
         return fs->strbufpos >= fs->strbufsize;
@@ -448,7 +246,7 @@ static int icvEof( CvFileStorage* fs )
     return false;
 }
 
-static void icvCloseFile( CvFileStorage* fs )
+void icvCloseFile( CvFileStorage* fs )
 {
     if( fs->file )
         fclose( fs->file );
@@ -463,7 +261,7 @@ static void icvCloseFile( CvFileStorage* fs )
     fs->is_opened = false;
 }
 
-static void icvRewind( CvFileStorage* fs )
+void icvRewind( CvFileStorage* fs )
 {
     if( fs->file )
         rewind(fs->file);
@@ -474,27 +272,6 @@ static void icvRewind( CvFileStorage* fs )
     fs->strbufpos = 0;
 }
 
-#define CV_YML_INDENT  3
-#define CV_XML_INDENT  2
-#define CV_YML_INDENT_FLOW  1
-#define CV_FS_MAX_LEN 4096
-
-#define CV_FILE_STORAGE ('Y' + ('A' << 8) + ('M' << 16) + ('L' << 24))
-#define CV_IS_FILE_STORAGE(fs) ((fs) != 0 && (fs)->flags == CV_FILE_STORAGE)
-
-#define CV_CHECK_FILE_STORAGE(fs)                       \
-{                                                       \
-    if( !CV_IS_FILE_STORAGE(fs) )                       \
-        CV_Error( (fs) ? CV_StsBadArg : CV_StsNullPtr,  \
-                  "Invalid pointer to file storage" );  \
-}
-
-#define CV_CHECK_OUTPUT_FILE_STORAGE(fs)                \
-{                                                       \
-    CV_CHECK_FILE_STORAGE(fs);                          \
-    if( !fs->write_mode )                               \
-        CV_Error( CV_StsError, "The file storage is opened for reading" ); \
-}
 
 CV_IMPL const char*
 cvAttrValue( const CvAttrList* attr, const char* attr_name )
@@ -629,8 +406,7 @@ inline char* icvFSResizeWriteBuffer( CvFileStorage* fs, char* ptr, int len )
 }
 
 
-static char*
-icvFSFlush( CvFileStorage* fs )
+char* icvFSFlush( CvFileStorage* fs )
 {
     char* ptr = fs->buffer;
     int indent;
@@ -1134,7 +910,7 @@ static bool is_param_exist( const std::vector<std::string> & params, const std::
     return std::find(params.begin(), params.end(), param) != params.end();
 }
 
-static void switch_to_Base64_state( CvFileStorage* fs, base64::fs::State state )
+void switch_to_Base64_state( CvFileStorage* fs, base64::fs::State state )
 {
     const char * err_unkonwn_state = "Unexpected error, unable to determine the Base64 state.";
     const char * err_unable_to_switch = "Unexpected error, unable to switch to this state.";
@@ -1197,7 +973,7 @@ static void switch_to_Base64_state( CvFileStorage* fs, base64::fs::State state )
 }
 
 
-static void check_if_write_struct_is_delayed( CvFileStorage* fs, bool change_type_to_base64 = false )
+void check_if_write_struct_is_delayed( CvFileStorage* fs, bool change_type_to_base64 )
 {
     if ( fs->is_write_struct_delayed )
     {
@@ -1346,8 +1122,6 @@ static void icvYMLGetMultilineStringContent(CvFileStorage* fs,
 
     end = ptr;
 }
-
-static int icvCalcStructSize( const char* dt, int initial_size );
 
 static char* icvYMLParseBase64(CvFileStorage* fs, char* ptr, int indent, CvFileNode * node)
 {
@@ -4751,8 +4525,7 @@ icvCalcElemSize( const char* dt, int initial_size )
 }
 
 
-static int
-icvCalcStructSize( const char* dt, int initial_size )
+int icvCalcStructSize( const char* dt, int initial_size )
 {
     int size = icvCalcElemSize( dt, initial_size );
     size_t elem_max_size = 0;
@@ -7479,935 +7252,11 @@ void read(const FileNode& node, std::string& value, const std::string& default_v
 
 
 
-/****************************************************************************
- * Newly added for Base64
- *
- *
- ***************************************************************************/
 
 
-/****************************************************************************
- * constant
- ***************************************************************************/
 
-#if CHAR_BIT != 8
-#error "`char` should be 8 bit."
-#endif
 
-base64::uint8_t const base64::base64_mapping[] =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-    "abcdefghijklmnopqrstuvwxyz"
-    "0123456789+/";
 
-base64::uint8_t const base64::base64_padding = '=';
 
-base64::uint8_t const base64::base64_demapping[] = {
-    0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,
-    0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,
-    0,  0,  0, 62,  0,  0,  0, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61,  0,  0,
-    0,  0,  0,  0,  0,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14,
-    15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25,  0,  0,  0,  0,  0,  0, 26, 27, 28,
-    29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48,
-    49, 50, 51,  0,  0,  0,  0,
-};
-
-/*    `base64_demapping` above is generated in this way:
- *    `````````````````````````````````````````````````````````````````````
- *  std::string mapping((const char *)base64_mapping);
- *    for (auto ch = 0; ch < 127; ch++) {
- *        auto i = mapping.find(ch);
- *        printf("%3u, ", (i != std::string::npos ? i : 0));
- *    }
- *    putchar('\n');
- *    `````````````````````````````````````````````````````````````````````
- */
-
-/****************************************************************************
- * function
- ***************************************************************************/
-
-size_t base64::base64_encode(uint8_t const * src, uint8_t * dst, size_t off, size_t cnt)
-{
-    if (!src || !dst || !cnt)
-        return 0;
-
-    /* initialize beginning and end */
-    uint8_t       * dst_beg = dst;
-    uint8_t       * dst_cur = dst_beg;
-
-    uint8_t const * src_beg = src + off;
-    uint8_t const * src_cur = src_beg;
-    uint8_t const * src_end = src_cur + cnt / 3U * 3U;
-
-    /* integer multiples part */
-    while (src_cur < src_end) {
-        uint8_t _2 = *src_cur++;
-        uint8_t _1 = *src_cur++;
-        uint8_t _0 = *src_cur++;
-        *dst_cur++ = base64_mapping[ _2          >> 2U];
-        *dst_cur++ = base64_mapping[(_1 & 0xF0U) >> 4U | (_2 & 0x03U) << 4U];
-        *dst_cur++ = base64_mapping[(_0 & 0xC0U) >> 6U | (_1 & 0x0FU) << 2U];
-        *dst_cur++ = base64_mapping[ _0 & 0x3FU];
-    }
-
-    /* remainder part */
-    size_t rst = src_beg + cnt - src_cur;
-    if (rst == 1U) {
-        uint8_t _2 = *src_cur++;
-        *dst_cur++ = base64_mapping[ _2          >> 2U];
-        *dst_cur++ = base64_mapping[(_2 & 0x03U) << 4U];
-    } else if (rst == 2U) {
-        uint8_t _2 = *src_cur++;
-        uint8_t _1 = *src_cur++;
-        *dst_cur++ = base64_mapping[ _2          >> 2U];
-        *dst_cur++ = base64_mapping[(_2 & 0x03U) << 4U | (_1 & 0xF0U) >> 4U];
-        *dst_cur++ = base64_mapping[(_1 & 0x0FU) << 2U];
-    }
-
-    /* padding */
-    switch (rst)
-    {
-    case 1U: *dst_cur++ = base64_padding;
-    case 2U: *dst_cur++ = base64_padding;
-    default: *dst_cur   = 0;
-        break;
-    }
-
-    return static_cast<size_t>(dst_cur - dst_beg);
-}
-
-size_t base64::base64_encode(char const * src, char * dst, size_t off, size_t cnt)
-{
-    if (cnt == 0U)
-        cnt = std::strlen(src);
-
-    return base64_encode
-    (
-        reinterpret_cast<uint8_t const *>(src),
-        reinterpret_cast<uint8_t       *>(dst),
-        off,
-        cnt
-    );
-}
-
-size_t base64::base64_decode(uint8_t const * src, uint8_t * dst, size_t off, size_t cnt)
-{
-    /* check parameters */
-    if (!src || !dst || !cnt)
-        return 0U;
-    if (cnt & 0x3U)
-        return 0U;
-
-    /* initialize beginning and end */
-    uint8_t       * dst_beg = dst;
-    uint8_t       * dst_cur = dst_beg;
-
-    uint8_t const * src_beg = src + off;
-    uint8_t const * src_cur = src_beg;
-    uint8_t const * src_end = src_cur + cnt;
-
-    /* start decoding */
-    while (src_cur < src_end) {
-        uint8_t d50 = base64_demapping[*src_cur++];
-        uint8_t c50 = base64_demapping[*src_cur++];
-        uint8_t b50 = base64_demapping[*src_cur++];
-        uint8_t a50 = base64_demapping[*src_cur++];
-
-        uint8_t b10 = b50 & 0x03U;
-        uint8_t b52 = b50 & 0x3CU;
-        uint8_t c30 = c50 & 0x0FU;
-        uint8_t c54 = c50 & 0x30U;
-
-        *dst_cur++ = (d50 << 2U) | (c54 >> 4U);
-        *dst_cur++ = (c30 << 4U) | (b52 >> 2U);
-        *dst_cur++ = (b10 << 6U) | (a50 >> 0U);
-    }
-
-    *dst_cur = 0;
-    return size_t(dst_cur - dst_beg);
-}
-
-size_t base64::base64_decode(char const * src, char * dst, size_t off, size_t cnt)
-{
-    if (cnt == 0U)
-        cnt = std::strlen(src);
-
-    return base64_decode
-    (
-        reinterpret_cast<uint8_t const *>(src),
-        reinterpret_cast<uint8_t       *>(dst),
-        off,
-        cnt
-    );
-}
-
-bool base64::base64_valid(uint8_t const * src, size_t off, size_t cnt)
-{
-    /* check parameters */
-    if (src == 0 || src + off == 0)
-        return false;
-    if (cnt == 0U)
-        cnt = std::strlen(reinterpret_cast<char const *>(src));
-    if (cnt == 0U)
-        return false;
-    if (cnt & 0x3U)
-        return false;
-
-    /* initialize beginning and end */
-    uint8_t const * beg = src + off;
-    uint8_t const * end = beg + cnt;
-
-    /* skip padding */
-    if (*(end - 1U) == base64_padding) {
-        end--;
-        if (*(end - 1U) == base64_padding)
-            end--;
-    }
-
-    /* find illegal characters */
-    for (uint8_t const * iter = beg; iter < end; iter++)
-        if (*iter > 126U || (!base64_demapping[(uint8_t)*iter] && *iter != base64_mapping[0]))
-            return false;
-
-    return true;
-}
-
-bool base64::base64_valid(char const * src, size_t off, size_t cnt)
-{
-    if (cnt == 0U)
-        cnt = std::strlen(src);
-
-    return base64_valid(reinterpret_cast<uint8_t const *>(src), off, cnt);
-}
-
-size_t base64::base64_encode_buffer_size(size_t cnt, bool is_end_with_zero)
-{
-    size_t additional = static_cast<size_t>(is_end_with_zero == true);
-    return (cnt + 2U) / 3U * 4U + additional;
-}
-
-size_t base64::base64_decode_buffer_size(size_t cnt, bool is_end_with_zero)
-{
-    size_t additional = static_cast<size_t>(is_end_with_zero == true);
-    return cnt / 4U * 3U + additional;
-}
-
-size_t base64::base64_decode_buffer_size(size_t cnt, char  const * src, bool is_end_with_zero)
-{
-    return base64_decode_buffer_size(cnt, reinterpret_cast<uchar const *>(src), is_end_with_zero);
-}
-
-size_t base64::base64_decode_buffer_size(size_t cnt, uchar const * src, bool is_end_with_zero)
-{
-    size_t padding_cnt = 0U;
-    for (uchar const * ptr = src + cnt - 1U; *ptr == base64_padding; ptr--)
-        padding_cnt ++;
-    return base64_decode_buffer_size(cnt, is_end_with_zero) - padding_cnt;
-}
-
-/****************************************************************************
- * to_binary && binary_to
- ***************************************************************************/
-
-template<typename _uint_t> inline size_t base64::
-to_binary(_uint_t val, uchar * cur)
-{
-    size_t delta = CHAR_BIT;
-    size_t cnt = sizeof(_uint_t);
-    while (cnt --> static_cast<size_t>(0U)) {
-        *cur++ = static_cast<uchar>(val);
-        val >>= delta;
-    }
-    return sizeof(_uint_t);
-}
-
-template<> inline size_t base64::to_binary(double val, uchar * cur)
-{
-    Cv64suf bit64;
-    bit64.f = val;
-    return to_binary(bit64.u, cur);
-}
-
-template<> inline size_t base64::to_binary(float val, uchar * cur)
-{
-    Cv32suf bit32;
-    bit32.f = val;
-    return to_binary(bit32.u, cur);
-}
-
-template<typename _primitive_t> inline size_t base64::
-to_binary(uchar const * val, uchar * cur)
-{
-    return to_binary<_primitive_t>(*reinterpret_cast<_primitive_t const *>(val), cur);
-}
-
-
-template<typename _uint_t> inline size_t base64::
-binary_to(uchar const * cur, _uint_t & val)
-{
-    val = static_cast<_uint_t>(0);
-    for (size_t i = static_cast<size_t>(0U); i < sizeof(_uint_t); i++)
-        val |= (static_cast<_uint_t>(*cur++) << (i * CHAR_BIT));
-    return sizeof(_uint_t);
-}
-
-template<> inline size_t base64::binary_to(uchar const * cur, double & val)
-{
-    Cv64suf bit64;
-    binary_to(cur, bit64.u);
-    val = bit64.f;
-    return sizeof(val);
-}
-
-template<> inline size_t base64::binary_to(uchar const * cur, float & val)
-{
-    Cv32suf bit32;
-    binary_to(cur, bit32.u);
-    val = bit32.f;
-    return sizeof(val);
-}
-
-template<typename _primitive_t> inline size_t base64::
-binary_to(uchar const * cur, uchar * val)
-{
-    return binary_to<_primitive_t>(cur, *reinterpret_cast<_primitive_t *>(val));
-}
-
-/****************************************************************************
- * others
- ***************************************************************************/
-
-std::string base64::make_base64_header(const char * dt)
-{
-    std::ostringstream oss;
-    oss << dt   << ' ';
-    std::string buffer(oss.str());
-    CV_Assert(buffer.size() < HEADER_SIZE);
-
-    buffer.reserve(HEADER_SIZE);
-    while (buffer.size() < HEADER_SIZE)
-        buffer += ' ';
-
-    return buffer;
-}
-
-bool base64::read_base64_header(std::vector<char> const & header, std::string & dt)
-{
-    std::istringstream iss(header.data());
-    return !!(iss >> dt);//the "std::basic_ios::operator bool" differs between C++98 and C++11. The "double not" syntax is portable and covers both cases with equivalent meaning
-}
-
-/****************************************************************************
- * Parser
- ***************************************************************************/
-
-base64::Base64ContextParser::Base64ContextParser(uchar * buffer, size_t size)
-    : dst_cur(buffer)
-    , dst_end(buffer + size)
-    , base64_buffer(BUFFER_LEN)
-    , src_beg(0)
-    , src_cur(0)
-    , src_end(0)
-    , binary_buffer(base64_encode_buffer_size(BUFFER_LEN))
-{
-    src_beg = binary_buffer.data();
-    src_cur = src_beg;
-    src_end = src_beg + BUFFER_LEN;
-}
-
-base64::Base64ContextParser::~Base64ContextParser()
-{
-    /* encode the rest binary data to base64 buffer */
-    if (src_cur != src_beg)
-        flush();
-}
-
-base64::Base64ContextParser & base64::Base64ContextParser::
-read(const uchar * beg, const uchar * end)
-{
-    if (beg >= end)
-        return *this;
-
-    while (beg < end) {
-        /* collect binary data and copy to binary buffer */
-        size_t len = std::min(end - beg, src_end - src_cur);
-        std::memcpy(src_cur, beg, len);
-        beg     += len;
-        src_cur += len;
-
-        if (src_cur >= src_end) {
-            /* binary buffer is full. */
-            /* decode it send result to dst */
-
-            CV_Assert(flush());    /* check for base64_valid */
-        }
-    }
-
-    return *this;
-}
-
-bool base64::Base64ContextParser::flush()
-{
-    if ( !base64_valid(src_beg, 0U, src_cur - src_beg) )
-        return false;
-
-    if ( src_cur == src_beg )
-        return true;
-
-    uchar * buffer = binary_buffer.data();
-    size_t len = base64_decode(src_beg, buffer, 0U, src_cur - src_beg);
-    src_cur = src_beg;
-
-    /* unexpected error */
-    CV_Assert(len != 0);
-
-    /* buffer is full */
-    CV_Assert(dst_cur + len < dst_end);
-
-    if (dst_cur + len < dst_end) {
-        /* send data to dst */
-        std::memcpy(dst_cur, buffer, len);
-        dst_cur += len;
-    }
-
-    return true;
-}
-
-/****************************************************************************
- * Emitter
- ***************************************************************************/
-
-/* A decorator for CvFileStorage
- * - no copyable
- * - not safe for now
- * - move constructor may be needed if C++11
- */
-class base64::Base64ContextEmitter
-{
-public:
-    explicit Base64ContextEmitter(CvFileStorage * fs)
-        : file_storage(fs)
-        , binary_buffer(BUFFER_LEN)
-        , base64_buffer(base64_encode_buffer_size(BUFFER_LEN))
-        , src_beg(0)
-        , src_cur(0)
-        , src_end(0)
-    {
-        src_beg = binary_buffer.data();
-        src_end = src_beg + BUFFER_LEN;
-        src_cur = src_beg;
-
-        CV_CHECK_OUTPUT_FILE_STORAGE(fs);
-
-        if ( fs->fmt == CV_STORAGE_FORMAT_JSON )
-        {
-            /* clean and break buffer */
-            *fs->buffer++ = '\0';
-            ::icvPuts( fs, fs->buffer_start );
-            fs->buffer = fs->buffer_start;
-            memset( file_storage->buffer_start, 0, static_cast<int>(file_storage->space) );
-            ::icvPuts( fs, "\"$base64$" );
-        }
-        else
-        {
-            ::icvFSFlush(file_storage);
-        }
-    }
-
-    ~Base64ContextEmitter()
-    {
-        /* cleaning */
-        if (src_cur != src_beg)
-            flush();    /* encode the rest binary data to base64 buffer */
-
-        if ( file_storage->fmt == CV_STORAGE_FORMAT_JSON )
-        {
-            /* clean and break buffer  */
-            ::icvPuts(file_storage, "\"");
-            file_storage->buffer = file_storage->buffer_start;
-            ::icvFSFlush( file_storage );
-            memset( file_storage->buffer_start, 0, static_cast<int>(file_storage->space) );
-            file_storage->buffer = file_storage->buffer_start;
-        }
-    }
-
-    Base64ContextEmitter & write(const uchar * beg, const uchar * end)
-    {
-        if (beg >= end)
-            return *this;
-
-        while (beg < end) {
-            /* collect binary data and copy to binary buffer */
-            size_t len = std::min(end - beg, src_end - src_cur);
-           std::memcpy(src_cur, beg, len);
-            beg     += len;
-            src_cur += len;
-
-            if (src_cur >= src_end) {
-                /* binary buffer is full. */
-                /* encode it to base64 and send result to fs */
-                flush();
-            }
-        }
-
-        return *this;
-    }
-
-    /*
-     * a convertor must provide :
-     * - `operator >> (uchar * & dst)` for writting current binary data to `dst` and moving to next data.
-     * - `operator bool` for checking if current loaction is valid and not the end.
-     */
-    template<typename _to_binary_convertor_t> inline
-    Base64ContextEmitter & write(_to_binary_convertor_t & convertor)
-    {
-        static const size_t BUFFER_MAX_LEN = 1024U;
-
-        std::vector<uchar> buffer(BUFFER_MAX_LEN);
-        uchar * beg = buffer.data();
-        uchar * end = beg;
-
-        while (convertor) {
-            convertor >> end;
-            write(beg, end);
-            end = beg;
-        }
-
-        return *this;
-    }
-
-    bool flush()
-    {
-        /* controll line width, so on. */
-        size_t len = base64_encode(src_beg, base64_buffer.data(), 0U, src_cur - src_beg);
-        if (len == 0U)
-            return false;
-
-        src_cur = src_beg;
-        {
-            if ( file_storage->fmt == CV_STORAGE_FORMAT_JSON )
-            {
-                ::icvPuts(file_storage, (const char*)base64_buffer.data());
-            }
-            else
-            {
-                const char newline[] = "\n";
-                char space[80];
-                int ident = file_storage->struct_indent;
-                memset(space, ' ', static_cast<int>(ident));
-                space[ident] = '\0';
-
-                ::icvPuts(file_storage, space);
-                ::icvPuts(file_storage, (const char*)base64_buffer.data());
-                ::icvPuts(file_storage, newline);
-                ::icvFSFlush(file_storage);
-            }
-
-        }
-
-        return true;
-    }
-
-private:
-    /* because of Base64, we must keep its length a multiple of 3 */
-    static const size_t BUFFER_LEN = 48U;
-    // static_assert(BUFFER_LEN % 3 == 0, "BUFFER_LEN is invalid");
-
-private:
-    CvFileStorage * file_storage;
-
-    std::vector<uchar> binary_buffer;
-    std::vector<uchar> base64_buffer;
-    uchar * src_beg;
-    uchar * src_cur;
-    uchar * src_end;
-};
-
-
-class base64::RawDataToBinaryConvertor
-{
-public:
-
-    RawDataToBinaryConvertor(const void* src, int len, const std::string & dt)
-        : beg(reinterpret_cast<const uchar *>(src))
-        , cur(0)
-        , end(0)
-    {
-        CV_Assert(src);
-        CV_Assert(!dt.empty());
-        CV_Assert(len > 0);
-
-        /* calc step and to_binary_funcs */
-        make_to_binary_funcs(dt);
-
-        end = beg;
-        cur = beg;
-
-        step = ::icvCalcStructSize(dt.c_str(), 0);
-        end = beg + step * static_cast<size_t>(len);
-    }
-
-    inline RawDataToBinaryConvertor & operator >>(uchar * & dst)
-    {
-        CV_DbgAssert(*this);
-
-        for (size_t i = 0U, n = to_binary_funcs.size(); i < n; i++) {
-            elem_to_binary_t & pack = to_binary_funcs[i];
-            pack.func(cur + pack.offset, dst + pack.offset);
-        }
-        cur += step;
-        dst += step;
-
-        return *this;
-    }
-
-    inline operator bool() const
-    {
-        return cur < end;
-    }
-
-private:
-    typedef size_t(*to_binary_t)(const uchar *, uchar *);
-    struct elem_to_binary_t
-    {
-        size_t      offset;
-        to_binary_t func;
-    };
-
-private:
-    void make_to_binary_funcs(const std::string &dt)
-    {
-        size_t cnt = 0;
-        size_t offset = 0;
-        char type = '\0';
-
-        std::istringstream iss(dt);
-        while (!iss.eof()) {
-            if (!(iss >> cnt)) {
-                iss.clear();
-                cnt = 1;
-            }
-            CV_Assert(cnt > 0U);
-            if (!(iss >> type))
-                break;
-
-            while (cnt-- > 0)
-            {
-                elem_to_binary_t pack;
-
-                size_t size = 0;
-                switch (type)
-                {
-                case 'u':
-                case 'c':
-                    size = sizeof(uchar);
-                    pack.func = to_binary<uchar>;
-                    break;
-                case 'w':
-                case 's':
-                    size = sizeof(ushort);
-                    pack.func = to_binary<ushort>;
-                    break;
-                case 'i':
-                    size = sizeof(uint);
-                    pack.func = to_binary<uint>;
-                    break;
-                case 'f':
-                    size = sizeof(float);
-                    pack.func = to_binary<float>;
-                    break;
-                case 'd':
-                    size = sizeof(double);
-                    pack.func = to_binary<double>;
-                    break;
-                case 'r':
-                default: { CV_Assert(!"type not support"); break; }
-                };
-
-                offset = static_cast<size_t>(cvAlign(static_cast<int>(offset), static_cast<int>(size)));
-                pack.offset = offset;
-                offset += size;
-
-                to_binary_funcs.push_back(pack);
-            }
-        }
-
-        CV_Assert(iss.eof());
-    }
-
-private:
-    const uchar * beg;
-    const uchar * cur;
-    const uchar * end;
-
-    size_t step;
-    std::vector<elem_to_binary_t> to_binary_funcs;
-};
-
-class base64::BinaryToCvSeqConvertor
-{
-public:
-    BinaryToCvSeqConvertor(const void* src, int len, const char* dt)
-        : cur(reinterpret_cast<const uchar *>(src))
-        , beg(reinterpret_cast<const uchar *>(src))
-        , end(reinterpret_cast<const uchar *>(src))
-    {
-        CV_Assert(src);
-        CV_Assert(dt);
-        CV_Assert(len >= 0);
-
-        /* calc binary_to_funcs */
-        make_funcs(dt);
-        functor_iter = binary_to_funcs.begin();
-
-        step = ::icvCalcStructSize(dt, 0);
-        end = beg + step * static_cast<size_t>(len);
-    }
-
-    inline BinaryToCvSeqConvertor & operator >> (CvFileNode & dst)
-    {
-        CV_DbgAssert(*this);
-
-        /* get current data */
-        union
-        {
-            uchar mem[sizeof(double)];
-            uchar  u;
-            char   b;
-            ushort w;
-            short  s;
-            int    i;
-            float  f;
-            double d;
-        } buffer; /* for GCC -Wstrict-aliasing */
-        std::memset(buffer.mem, 0, sizeof(buffer));
-        functor_iter->func(cur + functor_iter->offset, buffer.mem);
-
-        /* set node::data */
-        switch (functor_iter->cv_type)
-        {
-        case CV_8U : { dst.data.i = cv::saturate_cast<int>   (buffer.u); break;}
-        case CV_8S : { dst.data.i = cv::saturate_cast<int>   (buffer.b); break;}
-        case CV_16U: { dst.data.i = cv::saturate_cast<int>   (buffer.w); break;}
-        case CV_16S: { dst.data.i = cv::saturate_cast<int>   (buffer.s); break;}
-        case CV_32S: { dst.data.i = cv::saturate_cast<int>   (buffer.i); break;}
-        case CV_32F: { dst.data.f = cv::saturate_cast<double>(buffer.f); break;}
-        case CV_64F: { dst.data.f = cv::saturate_cast<double>(buffer.d); break;}
-        default: break;
-        }
-
-        /* set node::tag */
-        switch (functor_iter->cv_type)
-        {
-        case CV_8U :
-        case CV_8S :
-        case CV_16U:
-        case CV_16S:
-        case CV_32S: { dst.tag = CV_NODE_INT; /*std::printf("%i,", dst.data.i);*/ break; }
-        case CV_32F:
-        case CV_64F: { dst.tag = CV_NODE_REAL; /*std::printf("%.1f,", dst.data.f);*/ break; }
-        default: break;
-        }
-
-        /* check if end */
-        if (++functor_iter == binary_to_funcs.end()) {
-            functor_iter = binary_to_funcs.begin();
-            cur += step;
-        }
-
-        return *this;
-    }
-
-    inline operator bool() const
-    {
-        return cur < end;
-    }
-
-private:
-    typedef size_t(*binary_to_t)(uchar const *, uchar *);
-    struct binary_to_filenode_t
-    {
-        size_t      cv_type;
-        size_t      offset;
-        binary_to_t func;
-    };
-
-private:
-    void make_funcs(const char* dt)
-    {
-        size_t cnt = 0;
-        char type = '\0';
-        size_t offset = 0;
-
-        std::istringstream iss(dt);
-        while (!iss.eof()) {
-            if (!(iss >> cnt)) {
-                iss.clear();
-                cnt = 1;
-            }
-            CV_Assert(cnt > 0U);
-            if (!(iss >> type))
-                break;
-
-            while (cnt-- > 0)
-            {
-                binary_to_filenode_t pack;
-
-                /* set func and offset */
-                size_t size = 0;
-                switch (type)
-                {
-                case 'u':
-                case 'c':
-                    size      = sizeof(uchar);
-                    pack.func = binary_to<uchar>;
-                    break;
-                case 'w':
-                case 's':
-                    size      = sizeof(ushort);
-                    pack.func = binary_to<ushort>;
-                    break;
-                case 'i':
-                    size      = sizeof(uint);
-                    pack.func = binary_to<uint>;
-                    break;
-                case 'f':
-                    size      = sizeof(float);
-                    pack.func = binary_to<float>;
-                    break;
-                case 'd':
-                    size      = sizeof(double);
-                    pack.func = binary_to<double>;
-                    break;
-                case 'r':
-                default:  { CV_Assert(!"type not support"); break; }
-                }; // need a better way for outputting error.
-
-                offset = static_cast<size_t>(cvAlign(static_cast<int>(offset), static_cast<int>(size)));
-                pack.offset = offset;
-                offset += size;
-
-                /* set type */
-                switch (type)
-                {
-                case 'u': { pack.cv_type = CV_8U ; break; }
-                case 'c': { pack.cv_type = CV_8S ; break; }
-                case 'w': { pack.cv_type = CV_16U; break; }
-                case 's': { pack.cv_type = CV_16S; break; }
-                case 'i': { pack.cv_type = CV_32S; break; }
-                case 'f': { pack.cv_type = CV_32F; break; }
-                case 'd': { pack.cv_type = CV_64F; break; }
-                case 'r':
-                default:  { CV_Assert(!"type is not support"); break; }
-                } // need a better way for outputting error.
-
-                binary_to_funcs.push_back(pack);
-            }
-        }
-
-        CV_Assert(iss.eof());
-        CV_Assert(binary_to_funcs.size());
-    }
-
-private:
-
-    const uchar * cur;
-    const uchar * beg;
-    const uchar * end;
-
-    size_t step;
-    std::vector<binary_to_filenode_t> binary_to_funcs;
-    std::vector<binary_to_filenode_t>::iterator functor_iter;
-};
-
-
-
-/****************************************************************************
- * Wapper
- ***************************************************************************/
-
-
-base64::Base64Writer::Base64Writer(::CvFileStorage * fs)
-    : emitter(new Base64ContextEmitter(fs))
-    , data_type_string()
-{
-    CV_CHECK_OUTPUT_FILE_STORAGE(fs);
-}
-
-void base64::Base64Writer::write(const void* _data, size_t len, const char* dt)
-{
-    check_dt(dt);
-    RawDataToBinaryConvertor convertor(_data, static_cast<int>(len), data_type_string);
-    emitter->write(convertor);
-}
-
-template<typename _to_binary_convertor_t> inline
-void base64::Base64Writer::write(_to_binary_convertor_t & convertor, const char* dt)
-{
-    check_dt(dt);
-    emitter->write(convertor);
-}
-
-base64::Base64Writer::~Base64Writer()
-{
-    delete emitter;
-}
-
-void base64::Base64Writer::check_dt(const char* dt)
-{
-    if ( dt == 0 )
-        CV_Error( CV_StsBadArg, "Invalid \'dt\'." );
-    else if (data_type_string.empty()) {
-        data_type_string = dt;
-
-        /* output header */
-        std::string buffer = make_base64_header(dt);
-        const uchar * beg = reinterpret_cast<const uchar *>(buffer.data());
-        const uchar * end = beg + buffer.size();
-
-        emitter->write(beg, end);
-    } else if ( data_type_string != dt )
-        CV_Error( CV_StsBadArg, "\'dt\' does not match." );
-}
-
-
-void base64::make_seq(void * binary, int elem_cnt, const char * dt, ::CvSeq & seq)
-{
-    ::CvFileNode node;
-    node.info = 0;
-    BinaryToCvSeqConvertor convertor(binary, elem_cnt, dt);
-    while (convertor) {
-        convertor >> node;
-        cvSeqPush(&seq, &node);
-    }
-}
-
-void base64::cvWriteRawDataBase64(::CvFileStorage* fs, const void* _data, int len, const char* dt)
-{
-    CV_Assert(fs);
-    CV_CHECK_OUTPUT_FILE_STORAGE(fs);
-
-    check_if_write_struct_is_delayed( fs, true );
-
-    if ( fs->state_of_writing_base64 == base64::fs::Uncertain )
-    {
-        switch_to_Base64_state( fs, base64::fs::InUse );
-    }
-    else if ( fs->state_of_writing_base64 != base64::fs::InUse )
-    {
-        CV_Error( CV_StsError, "Base64 should not be used at present." );
-    }
-
-    fs->base64_writer->write(_data, len, dt);
-}
-
-/****************************************************************************
- * Interface
- ***************************************************************************/
-
-CV_IMPL void cvWriteRawDataBase64(::CvFileStorage* fs, const void* _data, int len, const char* dt)
-{
-    ::base64::cvWriteRawDataBase64(fs, _data, len, dt);
-}
 
 /* End of file. */
